Se separaron las clases por fichero y se agruparon en package relacionados
