+++ Projects_Java/Alpha/InventoryLogic/Inventory.java
@@ -1,16 +1,16 @@
-package Alpha;
-// Import necesario para trabajar con archivos CSV y Paths
-import java.nio.file.*; 
-// Import necesario para trabajar con archivos y flujos de entrada/salida de datos (I/O)
+package Alpha.InventoryLogic;
 import java.io.*; 
-// Import necesario para trabajar con colecciones y Scanner
-import java.util.*; 
+import java.util.*;
+
+import Alpha.Validation;
+import Alpha.Repositories.FileCSV;
+import Alpha.Repositories.InventoryRepository; 
 
 // Clase para controlar el inventario de productos y sus operaciones (agregar, modificar, eliminar, buscar, etc.)
-public class InventoryManagement {
+public class Inventory {
     private Map<String, List<Producto>> productosPorCategoria = new HashMap<>();
-    private alertaStock alerta = new alertaStock();
-    private ArchivoCSV archivoCSV = new ArchivoCSV();
+    private StockAlert alerta = new StockAlert();
+    private FileCSV ficherocsv = new FileCSV();
 
     public void agregarProducto(Producto producto) {
         productosPorCategoria.computeIfAbsent(producto.getCategoria(), k -> new ArrayList<>()).add(producto);
@@ -20,7 +20,7 @@ public class InventoryManagement {
         reporte.registrarCambio("Agregar Producto", producto.getCategoria(), producto.getNombre(), detalle);
     
         // Verificar stock bajo solo si es necesario
-        if (producto.getCantidad() < alertaStock.LIMITE_STOCK_BAJO) {
+        if (producto.getCantidad() < StockAlert.LIMITE_STOCK_BAJO) {
             alerta.verificarStockBajo(productosPorCategoria);
         }
     }
@@ -197,12 +197,12 @@ public class InventoryManagement {
         }
     }
 
-    public void cargarInventario(InventarioRepository repository) {
-        productosPorCategoria = repository.cargarInventario();
+    public void cargarInventario(InventoryRepository repositorioInventario) {
+        productosPorCategoria = repositorioInventario.cargarInventario();
     }
 
-    public void guardarInventario(InventarioRepository repository) {
-        repository.guardarInventario(productosPorCategoria);
+    public void guardarInventario(InventoryRepository repositorioInventario) {
+        repositorioInventario.guardarInventario(productosPorCategoria);
     }
 
     // Método para obtener los productos por categoría en el inventario
@@ -210,174 +210,12 @@ public class InventoryManagement {
         return productosPorCategoria;
     }
 
-    public void exportarInventario(File archivoCSV) {
-        this.archivoCSV.exportar(archivoCSV, productosPorCategoria);
+    public void exportarInventario(File ficherocsv) {
+        this.ficherocsv.exportar(ficherocsv, productosPorCategoria);
     }
 
-    public void importarInventario(File archivoCSV) {
-        this.archivoCSV.importar(archivoCSV, productosPorCategoria, this);
-    }
-}
-
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-
-class InventorySearch{
-
-}
-
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-
-class InventarioRepository {
-    private final File archivoInventario;
-
-    public InventarioRepository(File archivo) {
-        this.archivoInventario = new File("Projects_Java/Archivo_Proyecto_Alpha/.dats/Archivo_Inventario.dat");
-    }
-
-    @SuppressWarnings("unchecked")
-    public Map<String, List<Producto>> cargarInventario() {
-        Map<String, List<Producto>> productosPorCategoria = new HashMap<>();
-        try {
-            if (!archivoInventario.exists()) {
-                archivoInventario.getParentFile().mkdirs();
-                archivoInventario.createNewFile();
-                System.out.println("El archivo de inventario no existía. Se ha creado un nuevo archivo en: " + archivoInventario.getAbsolutePath());
-                return productosPorCategoria;
-            }
-
-            try (ObjectInputStream ois = new ObjectInputStream(new FileInputStream(archivoInventario))) {
-                Object obj = ois.readObject();
-                if (obj instanceof Map<?, ?>) {
-                    productosPorCategoria = (Map<String, List<Producto>>) obj;
-                } else {
-                    System.out.println("El archivo no contiene un formato válido de inventario.");
-                }
-            }
-        } catch (IOException | ClassNotFoundException e) {
-            System.out.println("Error al cargar el inventario: " + e.getMessage());
-        }
-        return productosPorCategoria;
-    }
-
-    public void guardarInventario(Map<String, List<Producto>> productosPorCategoria) {
-        try {
-            if (!archivoInventario.exists()) {
-                archivoInventario.getParentFile().mkdirs();
-                archivoInventario.createNewFile();
-            }
-
-            try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(archivoInventario))) {
-                oos.writeObject(productosPorCategoria);
-            }
-        } catch (IOException e) {
-            System.out.println("Error al guardar el inventario: " + e.getMessage());
-        }
+    public void importarInventario(File ficherocsv) {
+        this.ficherocsv.importar(ficherocsv, productosPorCategoria, this);
     }
 }
 
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-
-class ArchivoCSV {
-
-    public void exportar(File archivoCSV, Map<String, List<Producto>> productosPorCategoria) {
-        try (BufferedWriter writer = Files.newBufferedWriter(archivoCSV.toPath())) {
-            writer.write("Categoría,Producto,Cantidad,Precio\n"); // Encabezado del archivo CSV
-            for (var entry : productosPorCategoria.entrySet()) {
-                String categoria = entry.getKey();
-                for (Producto producto : entry.getValue()) {
-                    writer.write(String.format("%s,%s,%d,%.2f\n",
-                            categoria, producto.getNombre(), producto.getCantidad(), producto.getPrecio()));
-                }
-            }
-            System.out.println("Inventario exportado correctamente a " + archivoCSV.getAbsolutePath());
-        } catch (IOException e) {
-            System.out.println("Error al exportar el inventario a CSV: " + e.getMessage());
-        }
-    }
-
-    public void importar(File archivoCSV, Map<String, List<Producto>> productosPorCategoria, InventoryManagement controlStock) {
-        if (!archivoCSV.exists()) {
-            System.out.println("El archivo CSV especificado no existe en: " + archivoCSV.getAbsolutePath());
-            return;
-        }
-
-        try (BufferedReader reader = Files.newBufferedReader(archivoCSV.toPath())) {
-            String linea = reader.readLine(); // Leer encabezado
-            if (linea == null || !linea.equals("Categoría,Producto,Cantidad,Precio")) {
-                System.out.println("El archivo CSV no tiene el formato esperado. Asegúrate de que el encabezado sea 'Categoría,Producto,Cantidad,Precio'.");
-                return;
-            }
-
-            while ((linea = reader.readLine()) != null) {
-                String[] datos = linea.split(",");
-                if (datos.length == 4) {
-                    try {
-                        String categoria = datos[0].trim();
-                        String nombre = datos[1].trim();
-                        int cantidad = Integer.parseInt(datos[2].trim());
-                        double precio = Double.parseDouble(datos[3].trim());
-
-                        // Verificar si el producto ya existe
-                        Producto productoExistente = controlStock.buscarProducto(categoria, nombre);
-                        if (productoExistente != null) {
-                            // Actualizar cantidad y precio del producto existente
-                            productoExistente.setCantidad(productoExistente.getCantidad() + cantidad);
-                            productoExistente.setPrecio(precio);
-                            System.out.printf("Producto actualizado: Categoría='%s', Nombre='%s', Nueva Cantidad=%d, Nuevo Precio=%.2f\n",
-                                    categoria, nombre, productoExistente.getCantidad(), precio);
-                        } else {
-                            // Agregar nuevo producto
-                            Producto nuevoProducto = new Producto(nombre, categoria, cantidad, precio);
-                            controlStock.agregarProducto(nuevoProducto);
-                            System.out.printf("Producto importado: Categoría='%s', Nombre='%s', Cantidad=%d, Precio=%.2f\n",
-                                    categoria, nombre, cantidad, precio);
-                        }
-                    } catch (NumberFormatException e) {
-                        System.out.println("Error al procesar una línea del archivo CSV. Asegúrate de que los valores de cantidad y precio sean numéricos.");
-                    }
-                } else {
-                    System.out.println("Línea ignorada por formato incorrecto: " + linea);
-                }
-            }
-            System.out.println("Inventario importado correctamente desde " + archivoCSV.getAbsolutePath());
-        } catch (IOException e) {
-            System.out.println("Error al leer el archivo CSV: " + e.getMessage());
-        }
-    }
-}
-
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-/*--------------------------------------------------------------------------------------------------------------------------------------------------------*/
-
-// Clase para verificar y mostrar alertas de productos con stock bajo
-class alertaStock { 
-    public static final int LIMITE_STOCK_BAJO = 20; // Límite de stock bajo para mostrar la alerta
-
-    // Método para verificar si hay productos con stock bajo
-    public void verificarStockBajo(Map<String, List<Producto>> productosPorCategoria) { 
-        boolean hayStockBajo = false;
-        StringBuilder alerta = new StringBuilder("\n=== ALERTA: Productos con stock bajo ===\n");
-
-        for (var entry : productosPorCategoria.entrySet()) {
-            String categoria = entry.getKey();
-            for (Producto producto : entry.getValue()) {
-                if (producto.getCantidad() < LIMITE_STOCK_BAJO) {
-                    hayStockBajo = true;
-                    alerta.append(String.format("Categoría: %s | Producto: %s | Stock: %d | Precio: $%.2f\n",
-                            categoria, producto.getNombre(), producto.getCantidad(), producto.getPrecio()));
-                }
-            }
-        }
-
-        if (hayStockBajo) {
-            System.out.println(alerta);
-        }
-    }
-}
